requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openpyxl==3.1.5
